+++ package.config.cjs
@@ -1,13 +1,11 @@
-import {defineConfig} from '@sanity/pkg-utils'
+const {defineConfig} = require('@sanity/pkg-utils')
 
-export default defineConfig({
+module.exports = defineConfig({
   dist: 'dist',
   tsconfig: 'tsconfig.dist.json',
 
-  // Remove this block to enable strict export validation
   extract: {
     rules: {
-      'ae-forgotten-export': 'off',
       'ae-incompatible-release-tags': 'off',
       'ae-internal-missing-underscore': 'off',
       'ae-missing-release-tag': 'off',