+++ src/lib/iconManagerSetup.ts
@@ -1,12 +1,9 @@
-import {addAPIProvider, disableCache, enableCache} from '@iconify/react'
+import {addAPIProvider} from '@iconify/react'
 
 import IconManagerPluginOptions from '../types/IconManagerPluginOptions'
 
 const iconManagerSetup = (config: void | IconManagerPluginOptions): void => {
   try {
-    disableCache('all')
-    enableCache('session')
-
     if (config?.customEndpoint && new URL(config.customEndpoint)) {
       addAPIProvider('', {
         resources: [config.customEndpoint],