+++ src/styled.d.ts
@@ -1,0 +1,7 @@
+import 'styled-components'
+
+import type {Theme} from '@sanity/ui/theme'
+
+declare module 'styled-components' {
+  export interface DefaultTheme extends Theme {}
+}